add migration flyway V1__init_DB.sql and V2__add_admin.sql and corrected entities

+++ src/main/java/com/toolyard/toolyard_shop/model/OrderDetails.java
@@ -28,6 +28,6 @@ public class OrderDetails {
     @ManyToOne
     @JoinColumn(name = "product_id")
     private Product product;
-    private BigDecimal amount;
+    private int amount;
     private BigDecimal price;
 }
